+++ vllm/model_executor/layers/quantization/fp8.py
@@ -24,7 +24,7 @@ from vllm.model_executor.layers.quantization.base_config import (
     QuantizationConfig, QuantizeMethodBase)
 from vllm.model_executor.layers.quantization.kv_cache import BaseKVCacheMethod
 from vllm.model_executor.layers.quantization.utils.flashinfer_utils import (
-    FlashInferMoEBakcend, apply_flashinfer_per_tensor_scale_fp8,
+    FlashinferMoeBackend, apply_flashinfer_per_tensor_scale_fp8,
     build_flashinfer_fp8_cutlass_moe_kernel,
     flashinfer_fp8_cutlass_moe_forward, get_flashinfer_moe_backend,
     register_moe_scaling_factors,
@@ -491,11 +491,14 @@ class Fp8MoEMethod(FusedMoEMethodBase):
         self.quant_config = quant_config
         self.block_quant = self.quant_config.weight_block_size is not None
 
-        self.flashinfer_moe_backend: Optional[FlashInferMoEBakcend] = None
+        self.flashinfer_moe_backend: Optional[FlashinferMoeBackend] = None
         self.fused_experts: Optional[
             mk.FusedMoEModularKernel] = None  # type: ignore
         if envs.VLLM_USE_FLASHINFER_MOE_FP8 and has_flashinfer_moe():
             self.flashinfer_moe_backend = get_flashinfer_moe_backend()
+            logger.info_once(
+                f"Using FlashInfer {self.flashinfer_moe_backend.value} kernels"
+            )
         # For GPUs that lack FP8 hardware support, we can leverage the Marlin
         # kernel for fast weight-only FP8 quantization
         self.use_marlin = (not current_platform.has_device_capability(89)
@@ -536,7 +539,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                 "platform.")
 
     def maybe_swap_experts_impl(self, moe_parallel_config):
-        if self.flashinfer_moe_backend == FlashInferMoEBakcend.CUTLASS:
+        if self.flashinfer_moe_backend == FlashinferMoeBackend.CUTLASS:
             self.fused_experts = build_flashinfer_fp8_cutlass_moe_kernel(
                 moe_parallel_config)
 
@@ -850,7 +853,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                 register_moe_scaling_factors(layer)
                 w13_weight = swap_w13_to_w31(layer.w13_weight.data)
                 if self.flashinfer_moe_backend == \
-                    FlashInferMoEBakcend.FLASHINFER:
+                    FlashinferMoeBackend.FLASHINFER:
                     rotate_flashinfer_fp8_moe_weights(w13_weight, w2_weight)
                 layer.w13_weight.data = w13_weight.data
 
@@ -951,7 +954,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
             assert logical_replica_count is not None
             assert isinstance(layer, FusedMoE)
 
-        if self.flashinfer_moe_backend == FlashInferMoEBakcend.FLASHINFER:
+        if self.flashinfer_moe_backend == FlashinferMoeBackend.TENSORRT_LLM:
             assert activation == 'silu'
             assert scoring_func == 'sigmoid'
             if self.block_quant:
@@ -1047,7 +1050,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                 apply_router_weight_on_input=apply_router_weight_on_input,
                 global_num_experts=global_num_experts,
                 expert_map=expert_map)
-        elif self.flashinfer_moe_backend == FlashInferMoEBakcend.CUTLASS:
+        elif self.flashinfer_moe_backend == FlashinferMoeBackend.CUTLASS:
             assert self.fused_experts is not None
             assert self.block_quant is None
             assert (not renormalize and custom_routing_function is not None)
